requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
openai>=1.40.0
numpy>=1.26.0
pandas>=2.1.0
Pillow>=10.0.0
tqdm>=4.66.0
google-genai>=0.3.0
matplotlib>=3.8.0
